+++ src/tracing/instrumentation/ioredis.js
@@ -23,6 +23,7 @@ exports.init = function() {
 
 function instrument(ioredis) {
   shimmer.wrap(ioredis.prototype, 'sendCommand', instrumentSendCommand);
+  shimmer.wrap(ioredis.prototype, 'multi', instrumentMultiCommand);
   shimmer.wrap(ioredis.prototype, 'pipeline', instrumentPipelineCommand);
 }
 
@@ -39,97 +40,72 @@ function instrumentSendCommand(original) {
 
     var callback;
     var parentSpan = cls.getCurrentSpan();
-    if (cls.isExitSpan(parentSpan)) {
-      if (parentSpan.n === 'redis') {
-        // multi commands could actually be recorded as multiple spans, but we only want to record one
-        // batched span considering that a multi call represents a transaction.
-        // The same is true for pipeline calls, but they have a slightly different semantic.
-        var isMultiParent = parentSpan.data.redis.command === 'multi';
-        var isPipelineParent = parentSpan.data.redis.command === 'pipeline';
-        if (isMultiParent || isPipelineParent) {
-          var parentSpanSubCommands = parentSpan.data.redis.subCommands = parentSpan.data.redis.subCommands || [];
-          parentSpanSubCommands.push(command.name);
 
-          if (command.name.toLowerCase() === 'exec' &&
-            // pipelining is handled differently in ioredis
-            isMultiParent) {
-            // the exec call is actually when the transmission of these commands to redis is happening
-            parentSpan.ts = Date.now();
-            callback = cls.ns.bind(getMultiCommandEndCall(parentSpan));
-            command.promise.then(
-              // make sure that the first parameter is never truthy
-              callback.bind(null, null),
-              callback);
-          }
-        }
-      }
+    if (parentSpan.n === 'redis' &&
+       (parentSpan.data.redis.command === 'multi' ||
+        parentSpan.data.redis.command === 'pipeline') &&
+        // the multi call is handled in instrumentMultiCommand but since multi is also send to Redis it will also
+        // trigger instrumentSendCommand, which is why we filter it out.
+        command.name !== 'multi') {
+      // multi commands could actually be recorded as multiple spans, but we only want to record one
+      // batched span considering that a multi call represents a transaction.
+      // The same is true for pipeline calls, but they have a slightly different semantic.
+      var parentSpanSubCommands = parentSpan.data.redis.subCommands = parentSpan.data.redis.subCommands || [];
+      parentSpanSubCommands.push(command.name);
+    } else if (cls.isExitSpan(parentSpan)) {
+      // Apart from the special case of multi/pipeline calls, redis exits can't be child spans of other exits.
       return original.apply(this, arguments);
     }
 
-    var span = cls.startSpan('redis');
-    span.stack = tracingUtil.getStackTrace(wrappedInternalSendCommand);
-    span.data = {
-      redis: {
-        connection: client.options.host + ':' + client.options.port,
-        command: command.name.toLowerCase()
+    var argsForOriginal = arguments;
+    return cls.ns.runAndReturn(function() {
+      var span = cls.startSpan('redis');
+      span.stack = tracingUtil.getStackTrace(wrappedInternalSendCommand);
+      span.data = {
+        redis: {
+          connection: client.options.host + ':' + client.options.port,
+          command: command.name.toLowerCase()
+        }
+      };
+
+      callback = cls.ns.bind(onResult);
+      command.promise.then(
+        // make sure that the first parameter is never truthy
+        callback.bind(null, null),
+        callback);
+
+      return original.apply(client, argsForOriginal);
+
+      function onResult(error) {
+        // multi commands are ended by exec. Wait for the exec result
+        if (command.name === 'multi') {
+          return;
+        }
+
+        span.d = Date.now() - span.ts;
+
+        if (error) {
+          span.error = true;
+          span.ec = 1;
+          span.data.redis.error = error.message;
+        }
+
+        span.transmit();
       }
-    };
-
-    callback = cls.ns.bind(onResult);
-    command.promise.then(
-      // make sure that the first parameter is never truthy
-      callback.bind(null, null),
-      callback);
-    return original.apply(this, arguments);
-
-    function onResult(error) {
-      // multi commands are ended by exec. Wait for the exec result
-      if (command.name === 'multi') {
-        return;
-      }
-
-      span.d = Date.now() - span.ts;
-
-      if (error) {
-        span.error = true;
-        span.ec = 1;
-        span.data.redis.error = error.message;
-      }
-
-      span.transmit();
-    }
+    });
   };
 }
 
-function getMultiCommandEndCall(span) {
-  return function multiCommandEndCallback(error) {
-    span.d = Date.now() - span.ts;
-
-    var subCommands = span.data.redis.subCommands;
-    var commandCount = 1;
-    if (subCommands) {
-      // remove exec call
-      subCommands.pop();
-      commandCount = subCommands.length;
-    }
-
-    span.b = {
-      s: commandCount,
-      u: false
-    };
-
-    if (error) {
-      span.error = true;
-      span.ec = commandCount;
-      span.data.redis.error = error.message;
-    }
-
-    span.transmit();
-  };
+function instrumentMultiCommand(original) {
+  return instrumentMultiOrPipelineCommand('multi', original);
 }
 
 function instrumentPipelineCommand(original) {
-  return function wrappedInternalPipelineCommand() {
+  return instrumentMultiOrPipelineCommand('pipeline', original);
+}
+
+function instrumentMultiOrPipelineCommand(commandName, original) {
+  return function wrappedInternalMultiOrPipelineCommand() {
     var client = this;
 
     if (!isActive ||
@@ -142,39 +118,77 @@ function instrumentPipelineCommand(original) {
       return original.apply(this, arguments);
     }
 
+    // create a new cls context parent to track the multi/pipeline child calls
+    var clsContextForMultiOrPipeline = cls.ns.createContext();
+    cls.ns.enter(clsContextForMultiOrPipeline);
     var span = cls.startSpan('redis');
-    span.stack = tracingUtil.getStackTrace(wrappedInternalPipelineCommand);
+    span.stack = tracingUtil.getStackTrace(wrappedInternalMultiOrPipelineCommand);
     span.data = {
       redis: {
         connection: client.options.host + ':' + client.options.port,
-        command: 'pipeline'
+        command: commandName
       }
     };
 
-    var pipeline = original.apply(this, arguments);
-    shimmer.wrap(pipeline, 'exec', instrumentPipelineExec.bind(null, span));
-    return pipeline;
+    var multiOrPipeline = original.apply(this, arguments);
+    shimmer.wrap(
+      multiOrPipeline,
+      'exec',
+      instrumentMultiOrPipelineExec.bind(null, clsContextForMultiOrPipeline, commandName, span)
+    );
+    return multiOrPipeline;
   };
 }
 
-function instrumentPipelineExec(span, original) {
-  return function instrumentedPipelineExec() {
-    // the exec call is actually when the transmission of these commands to redis is happening
+function instrumentMultiOrPipelineExec(clsContextForMultiOrPipeline, commandName, span, original) {
+  var endCallback =
+    commandName === 'pipeline' ?
+      pipelineCommandEndCallback :
+      multiCommandEndCallback;
+  return function instrumentedExec() {
+    // the exec call is actually when the transmission of these commands to
+    // redis is happening
     span.ts = Date.now();
 
     var result = original.apply(this, arguments);
     if (result.then) {
       result.then(function(results) {
-        pipelineCommandEndCallback(span, null, results);
+        endCallback.call(null, clsContextForMultiOrPipeline, span, null, results);
       }, function(error) {
-        pipelineCommandEndCallback(span, error, []);
+        endCallback.call(null, clsContextForMultiOrPipeline, span, error, []);
       });
     }
     return result;
   };
 }
 
-function pipelineCommandEndCallback(span, error, results) {
+function multiCommandEndCallback(clsContextForMultiOrPipeline, span, error) {
+  span.d = Date.now() - span.ts;
+
+  var subCommands = span.data.redis.subCommands;
+  var commandCount = 1;
+  if (subCommands) {
+    // remove exec call
+    subCommands.pop();
+    commandCount = subCommands.length;
+  }
+
+  span.b = {
+    s: commandCount,
+    u: false
+  };
+
+  if (error) {
+    span.error = true;
+    span.ec = commandCount;
+    span.data.redis.error = error.message;
+  }
+
+  span.transmit();
+  cls.ns.exit(clsContextForMultiOrPipeline);
+}
+
+function pipelineCommandEndCallback(clsContextForMultiOrPipeline, span, error, results) {
   span.d = Date.now() - span.ts;
 
   var subCommands = span.data.redis.subCommands;
@@ -211,4 +225,5 @@ function pipelineCommandEndCallback(span, error, results) {
   }
 
   span.transmit();
+  cls.ns.exit(clsContextForMultiOrPipeline);
 }