+++ test/tracing/httpClient/test.js
@@ -1,0 +1,118 @@
+'use strict';
+
+var expect = require('chai').expect;
+
+var supportedVersion = require('../../../src/tracing/index').supportedVersion;
+var config = require('../../config');
+var utils = require('../../utils');
+
+describe('tracing/httpClient', function() {
+  if (!supportedVersion(process.versions.node)) {
+    return;
+  }
+
+  // controls require features that aren't available in early Node.js versions
+  var agentControls = require('../../apps/agentStubControls');
+  var ClientControls = require('./clientControls');
+  var ServerControls = require('./serverControls');
+
+  this.timeout(config.getTestTimeout());
+
+  agentControls.registerTestHooks({
+    extraHeaders: ['fooBaR']
+  });
+
+  var serverControls = new ServerControls({
+    agentControls: agentControls
+  });
+  serverControls.registerTestHooks();
+
+  var clientControls = new ClientControls({
+    agentControls: agentControls,
+    env: {
+      SERVER_PORT: serverControls.port
+    }
+  });
+  clientControls.registerTestHooks();
+
+  it('must trace calls that fail due to connection refusal', function() {
+    return serverControls.kill()
+      .then(function() {
+        return clientControls.sendRequest({
+          method: 'GET',
+          path: '/timeout',
+          simple: false
+        });
+      })
+      .then(function() {
+        return utils.retry(function() {
+          return agentControls.getSpans()
+          .then(function(spans) {
+            utils.expectOneMatching(spans, function(span) {
+              expect(span.n).to.equal('node.http.client');
+              expect(span.ec).to.equal(1);
+              expect(span.data.http.error).to.match(/ECONNREFUSED/);
+            });
+          });
+        });
+      });
+  });
+
+  it('must trace calls that fail due to timeouts', function() {
+    return clientControls.sendRequest({
+      method: 'GET',
+      path: '/timeout',
+      simple: false
+    })
+      .then(function() {
+        return utils.retry(function() {
+          return agentControls.getSpans()
+          .then(function(spans) {
+            utils.expectOneMatching(spans, function(span) {
+              expect(span.n).to.equal('node.http.client');
+              expect(span.ec).to.equal(1);
+              expect(span.data.http.error).to.match(/Timeout/);
+            });
+          });
+        });
+      });
+  });
+
+  it('must trace aborted calls', function() {
+    return clientControls.sendRequest({
+      method: 'GET',
+      path: '/abort',
+      simple: false
+    })
+      .then(function() {
+        return utils.retry(function() {
+          return agentControls.getSpans()
+          .then(function(spans) {
+            utils.expectOneMatching(spans, function(span) {
+              expect(span.n).to.equal('node.http.client');
+              expect(span.ec).to.equal(1);
+              expect(span.data.http.error).to.match(/aborted/);
+            });
+          });
+        });
+      });
+  });
+
+  it('must record custom headers', function() {
+    return clientControls.sendRequest({
+      method: 'GET',
+      path: '/'
+    })
+      .then(function() {
+        return utils.retry(function() {
+          return agentControls.getSpans()
+          .then(function(spans) {
+            utils.expectOneMatching(spans, function(span) {
+              expect(span.n).to.equal('node.http.client');
+              expect(span.data.http.header.foobar).to.equal('42');
+            });
+          });
+        });
+      });
+  });
+});